Fix repping in threads of channels which reps shouldn't be allowed in (#78)
Threads have their own unique IDs and do not have the same ID as the channel which is why repping is possible in threads in channels like off-topic, anon-confessions etc. This fix checks if the channel is a thread and gets the parent ID of the channel if it is a thread.